Added getAllPointsInRadius to utilities.js and associated tests
getAllPointsInRadius is a function that will return all points which
are within a given radius from a central point. Associated tests
were added as well. Because of the way utilities.js is added into
the main project, the export necessary to test the script is commented
out. Uncomment the exports at the bottom of the file before testing.