Update links to contributing and code of conduct
GitHub normally provides relative linking for repo files:
https://help.github.com/articles/basic-writing-and-formatting-syntax

However, when these relative links are included within issues or PRs,
they lead to the "Compare changes" screen instead of the file itself.

I changed the links to absolute URLs to the docs in the dev branch.

When we merge dev into master, we can point the links to master.